refactor: tasks table placed in separate component
