+++ src/app/components/table-task/table-task.component.html
@@ -1,0 +1,43 @@
+<table class="table table-striped table-sm">
+  <thead>
+      <tr>
+          @for (column of tableColumn; track column) {
+              <th>
+                  {{column.title}}
+                  <button>
+                      <mat-icon>filter_list</mat-icon>
+                  </button>
+                  <button (click)="sortBy(column.name)">
+                      @if (sortProperty === column.name) {
+                          @switch (sortOrder) {
+                              @case (-1) {
+                                  <mat-icon>sort</mat-icon>
+                              }
+                              @case (1) {
+                                  <mat-icon style="transform: scale(1, -1)">sort</mat-icon>
+                              }
+                              @case (0) {
+                                  <mat-icon>sort_by_alpha</mat-icon>
+                              }
+                          }
+                      } @else {
+                          <mat-icon>sort_by_alpha</mat-icon>
+                      }
+                  </button>
+              </th>
+          }
+      </tr>
+  </thead>
+  <tbody>
+    @for (task of tasks; track task.id) {
+    <tr>
+        <td>{{task.ended ? 'Выполнено': 'Не выполнено'}}</td>
+        <td>{{task.title}}</td>
+        <td>{{task.beginDate | date: "dd.MM.YYYY":"":"ru"}}</td>
+        <td>{{task.deadline | date: "dd.MM.YYYY":"":"ru"}}</td>
+        <td>{{task.priority}}</td>
+        <td>{{task.status}}</td>
+    </tr>
+    }
+  </tbody>
+</table>
